Create featrank_glmnet_upd.R
Updated lasso feature ranking function with additional options. run.type = ("cv","nocv") - use cv.glmnet or glmnet. ranking.type = (1,2) - rank features by the number of lambda values where the coefficients are non-zero or rank features by the order in which they permanently drop to zero. The two types of ranking differ only when feature coefficients do not descend to zero monotonically.
Defaults weights are set to a column of 1 because glmnet (but not cv.glmnet) does not accept weights = NULL explicitly passed as an argument.